requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.1.1
python-dotenv==1.0.0
email-validator==2.0.0
pytest==7.4.0
ruff==0.0.290
mypy==1.5.1
